# Bug-fixed and optimized lowest common ancestor