--- a/skbio/tree/_tree.py
+++ b/skbio/tree/_tree.py
@@ -546,7 +546,7 @@ class TreeNode(SkbioObject):
             return [n for n in nodes if n is not ignore]
 
     def lowest_common_ancestor(self, tipnames):
-        r"""Find the lowest common ancestor of a list of tips.
+        r"""Find the lowest common ancestor of a list of nodes.
 
         Parameters
         ----------
@@ -563,6 +563,11 @@ class TreeNode(SkbioObject):
         ValueError
             If no tips could be found in the tree, or if not all tips were found.
 
+        Notes
+        -----
+        Despite the parameter is named as ``tipnames``, it can accept both tips and
+        internal nodes.
+
         Examples
         --------
         >>> from skbio import TreeNode
@@ -577,41 +582,38 @@ class TreeNode(SkbioObject):
         root
 
         """
-        if len(tipnames) == 1:
-            return self.find(next(iter(tipnames)))
+        nodes = [self.find(x) for x in tipnames]
+        if not nodes:
+            raise ValueError("No node is found.")
+        elif len(nodes) == 1:
+            return nodes[0]
 
-        tips = [self.find(name) for name in tipnames]
+        # A temporary attribute "prev" will be assigned to visited nodes. It represents
+        # the previous node in the upward path, or None, which indicates the current
+        # node is the beginning of the path, or there are more than one previous node.
+        visited = []
+        visited_append = visited.append
 
-        if len(tips) == 0:
-            raise ValueError("No tips found.")
-
-        nodes_to_scrub = []
-
-        for t in tips:
-            if t.is_root():
-                # has to be the LCA...
-                return t
-
-            prev = t
-            curr = t.parent
-
-            while curr and not hasattr(curr, "black"):
-                setattr(curr, "black", [prev])
-                nodes_to_scrub.append(curr)
+        for curr in nodes:
+            prev = None
+            while curr is not None:
+                # set prev as None if already visited
+                if hasattr(curr, "prev"):
+                    curr.prev = None
+                    break
+                curr.prev = prev
+                visited_append(curr)
                 prev = curr
                 curr = curr.parent
 
-            # increase black count, multiple children lead to here
-            if curr:
-                curr.black.append(prev)
-
+        # walk down the tree until last node with prev is None
         curr = self
-        while len(curr.black) == 1:
-            curr = curr.black[0]
+        while (prev := curr.prev) is not None:
+            curr = prev
 
-        # clean up tree
-        for n in nodes_to_scrub:
-            delattr(n, "black")
+        # clean up temporary attribute "prev"
+        for node in visited:
+            delattr(node, "prev")
 
         return curr