--- a/skbio/tree/tests/test_tree.py
+++ b/skbio/tree/tests/test_tree.py
@@ -236,27 +236,32 @@ class TreeTests(TestCase):
         self.assertEqual(obs, exp)
 
     def test_lowest_common_ancestor(self):
-        """TreeNode lowestCommonAncestor should return LCA for set of tips"""
+        """Return LCA for set of nodes."""
         t1 = TreeNode.read(["((a,(b,c)d)e,f,(g,h)i)j;"])
         t2 = t1.copy()
         t3 = t1.copy()
         t4 = t1.copy()
+        t5 = t1.copy()
         input1 = ["a"]  # return self
         input2 = ["a", "b"]  # return e
         input3 = ["b", "c"]  # return d
         input4 = ["a", "h", "g"]  # return j
+        input5 = ["g", "i"]  # return i (nested)
         exp1 = t1.find("a")
         exp2 = t2.find("e")
         exp3 = t3.find("d")
         exp4 = t4
+        exp5 = t5.find("i")
         obs1 = t1.lowest_common_ancestor(input1)
         obs2 = t2.lowest_common_ancestor(input2)
         obs3 = t3.lowest_common_ancestor(input3)
         obs4 = t4.lowest_common_ancestor(input4)
+        obs5 = t5.lowest_common_ancestor(input5)
         self.assertEqual(obs1, exp1)
         self.assertEqual(obs2, exp2)
         self.assertEqual(obs3, exp3)
         self.assertEqual(obs4, exp4)
+        self.assertEqual(obs5, exp5)
 
         # verify multiple calls work
         t_mul = t1.copy()
@@ -266,6 +271,12 @@ class TreeTests(TestCase):
         obs_2 = t_mul.lowest_common_ancestor(["g", "h"])
         self.assertEqual(obs_1, exp_1)
         self.assertEqual(obs_2, exp_2)
+
+        # verify subtree call
+        t_sub = t1.copy()
+        obs = t_sub.find("e").lowest_common_ancestor(["b", "c"])
+        exp = t_sub.find("d")
+        self.assertEqual(obs, exp)
 
         # root included
         t_root = TreeNode.read(["(a,b)c;"])