This PR fixed a bug in the [lowest common ancestor](https://en.wikipedia.org/wiki/Lowest_common_ancestor) algorithm of the tree structure. If one of the input nodes is ancestral to other nodes, the `lca` method will return the parent of that node, which is incorrect, instead of the node itself. For example:

```python
>>> tree = TreeNode.read(['(((a,b)c,d)e,(f,(g,h)i)j)k;'])
>>> print(tree.ascii_art())
                              /-a
                    /c-------|
          /e-------|          \-b
         |         |
-k-------|          \-d
         |
         |          /-f
          \j-------|
                   |          /-g
                    \i-------|
                              \-h
>>> tree.lca(['a', 'b', 'c']).name  # the correct result should be 'c'
'e'
```

This behavior is unlikely to be intensional, because supplying a single node will return the node itself.

```python
>>> tree.lca(['c']).name
'c'
```

This PR fixed this bug, and also improved the performance of the LCA algorithm. Benchmark:

Before: 1.78 ms ± 205 µs
After: 1.2 ms ± 133 µs

***

Please complete the following checklist:

* [x] I have read the [contribution guidelines](https://scikit.bio/contribute.html).

* [x] I have documented all public-facing changes in the [changelog](https://github.com/scikit-bio/scikit-bio/blob/main/CHANGELOG.md).

* [ ] **This pull request includes code, documentation, or other content derived from external source(s).** If this is the case, ensure the external source's license is compatible with scikit-bio's [license](https://github.com/scikit-bio/scikit-bio/blob/main/LICENSE.txt). Include the license in the `licenses` directory and add a comment in the code giving proper attribution. Ensure any other requirements set forth by the license and/or author are satisfied.

  - **It is your responsibility to disclose** code, documentation, or other content derived from external source(s). If you have questions about whether something can be included in the project or how to give proper attribution, include those questions in your pull request and a reviewer will assist you.

* [x] This pull request does not include code, documentation, or other content derived from external source(s).

Note: [This document](https://scikit.bio/devdoc/review.html) may also be helpful to see some of the things code reviewers will be verifying when reviewing your pull request.